added scripts to train on condor with singularity

+++ scripts/set_up_vqa_training.sh
@@ -1,7 +1,14 @@
 USERNAME=$1
 OUTPUT_DIR=$2
 
+cp scripts/run_train_singularity.sh /scratch/cluster/$USERNAME/uniter_image/
+
 cd /scratch/cluster/$USERNAME/uniter_image
+
+mkdir -p uniter
+mkdir -p vqa_dataset
 mkdir -p $OUTPUT_DIR
 mkdir -p $OUTPUT_DIR/ckpt
-cp vqa_dataset/pretrained/uniter-base.pt $OUTPUT_DIR/ckpt/model_step_0.pt
+if [ ! -f  $OUTPUT_DIR/ckpt/model_step_0.pt]
+  cp vqa_dataset/pretrained/uniter-base.pt $OUTPUT_DIR/ckpt/model_step_0.pt
+fi
